+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Change Log
 
+### Unreleased
+
 ### [v6.126](https://github.com/replicant4j/replicant/tree/v6.126) (2022-09-27) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.125...v6.126)
 
 Changes in this release: